backend/requirements.txt
# FastAPI core
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1

# Validation
pydantic[email]>=2.5.0

# Config
python-dotenv>=1.0.0

# AI (Phase 3 will use these)
groq>=0.11.0
pdfplumber>=0.10.0
pytesseract>=0.3.10
pdf2image>=1.17.0
Pillow>=10.0.0
